Extract tests stubs into standalone class modules
